Adds the `<@>` and `<@` operators.
This is useful for combining several Behaviors or Dynamics
together and sampling them with an Event.

There is an example in the Haddocks for the typeclass that
might help explain further / motivate the addition.

These operators are defined for `Behavior` but not `Dynamic`.
This removes the need for a typeclass, although it can be
added in later if people identify a need for it.