Rename Data.{Fast,}WeakBag.traverse to Data.{Fast,}WeakBag.traverse_
Renamed Data.FastWeakBag.traverse and Data.WeakBag.traverse to
Data.FastWeakBag.traverse_ and Data.WeakBag.traverse_ respectively.
This is a more accurate naming for the function as it has no resulting
value in the same way Data.Foldable.traverse_, and unlike
Data.Traverseable.traverse which has a resulting value. In addition,
we do not have to ignore any hlint warnings any more.