Fix TH splicing builds in CI
`haskell-overlays/splices-load-save/load-splices.nix` in reflex-platform uses
the name of the package to pull out the splices. But we have two packages both
with the package name "reflex". This hack makes the TH one have the `reflex`
attribute name, so we splice with TH.

Gross, but gets the job done.