Remove coincidence and add occurs (incorrect implementation).
The idea would be to have a dual to `now`, called `occurs` with type
`Event t a -> m (Maybe a)`. The `coincidence` function is then simply
a derived function equal to `push occurs`.

My first attempt at implementing `occurs` in Spider is obviously not
correct since many tests concerning `coincidence` are failing. If
`coincidence` has such an involved implementation then `occurs`
probably will too, and it might also be that `coincidence` needs to exist
as a separate function for efficiency reasons?